# swagger.yml
openapi: 3.0.0
info:
  title: Event Locator API
  version: 1.0.0
  description: |
    An API for our toy city in Kigali, Rwanda, to manage citizens and events, with location-based notifications delivered by the town crier!
    Supports multiple languages (en, es, fr) via the `Accept-Language` header or `lang` query parameter. Example: `Accept-Language: fr` or `?lang=fr` for French responses.
servers:
  - url: http://localhost:3000
    description: Local development server in Kigali
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Event:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: Music Night
        description:
          type: string
          example: A night of music
        location:
          type: object
          properties:
            type:
              type: string
              example: Point
            coordinates:
              type: array
              items:
                type: number
              example: [30.0928, -1.9482] # [lng, lat]
        date_time:
          type: string
          format: date-time
          example: 2025-04-10T18:00:00Z
        preferences:
          type: array
          items:
            type: string
          example: ["music"]
        creator_id:
          type: integer
          example: 1
...
paths:
...
  /events:
    post:
      summary: Create New Event
      description: Creates a new event in Kigali’s toy city. Requires authentication.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: lang
          schema:
            type: string
            enum: ["en", "es", "fr"]
            default: "en"
          description: Optional language override (e.g., 'fr' for French).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: Music Night
                  description: Title of the event
                description:
                  type: string
                  example: A night of music
                  description: Description of the event
                location:
                  type: string
                  example: "Kigali, Rwanda"
                  description: Event location (converted to coordinates)
                date_time:
                  type: string
                  format: date-time
                  example: 2025-04-10T18:00:00Z
                  description: Date and time of the event
                preferences:
                  type: array
                  items:
                    type: string
                  example: ["music"]
                  description: Event preferences or categories
              required:
                - title
                - description
                - location
                - date_time
                - preferences
      responses:
        "201":
          description: Event created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Event created successfully" # en
                    examples:
                      en: "Event created successfully"
                      es: "Evento creado con éxito"
                      fr: "Événement créé avec succès"
                  id:
                    type: integer
                    example: 1
                  title:
                    type: string
                    example: Music Night
        "401":
          description: Unauthorized (no valid token)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Unauthorized" # en
                    examples:
                      en: "Unauthorized"
                      es: "No autorizado"
                      fr: "Non autorisé"
        "500":
          description: Server error (e.g., geocoding failure)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal server error" # en
                    examples:
                      en: "Internal server error"
                      es: "Error interno del servidor"
                      fr: "Erreur interne du serveur"

    get:
      summary: Get All Events
      description: Retrieves all events in Kigali’s toy city.
      parameters:
        - in: query
          name: lang
          schema:
            type: string
            enum: ["en", "es", "fr"]
            default: "en"
          description: Optional language override (e.g., 'fr' for French).
      responses:
        "200":
          description: Events retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Events retrieved successfully" # en
                    examples:
                      en: "Events retrieved successfully"
                      es: "Eventos recuperados con éxito"
                      fr: "Événements récupérés avec succès"
                  events:
                    type: array
                    items:
                      $ref: "#/components/schemas/Event"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal server error" # en
                    examples:
                      en: "Internal server error"
                      es: "Error interno del servidor"
                      fr: "Erreur interne du serveur"
...